+++ src/adm/precache.rs
@@ -1,6 +1,6 @@
 use super::result::Error;
 use crate::models::{Photo, SizeTag};
-use crate::photosdir::{get_scaled_jpeg, PhotosDir};
+use crate::photosdir::{PhotosDir, get_scaled_jpeg};
 use crate::schema::photos::dsl::{date, is_public};
 use crate::{CacheOpt, DbOpt, DirOpt};
 use diesel::prelude::*;