+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on
 
 ## Unreleased
 
+* Use rust 1.85 and edition 2024.
 * Updated dependeices to `brotli` 7.0, `image` 0.24.9,
   `kamadak-exif` 0.6.1, `rand` 0.9, and `ructe` 0.18.2.
 * Some more known kinds of places.