+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on
 * Use rust 1.85 and edition 2024.
 * Updated dependeices to `brotli` 7.0, `image` 0.24.9,
   `kamadak-exif` 0.6.1, `rand` 0.9, and `ructe` 0.18.2.
+* Removed `async_trait` dependecy.
 * Some more known kinds of places.